trying to add comment likes

+++ src/main/java/com/makersacademy/acebook/model/CommentLike.java
@@ -24,7 +24,7 @@ public class CommentLike {
     @Column(name = "like_timestamp")
     private Timestamp timestamp;
 
-    public CommentLike(int userId, int commentId){
+    public CommentLike(long userId, long commentId){
         this.userId = userId;
         this.commentId = commentId;
         this.timestamp = Timestamp.from(Instant.now());
